Silhouette (Clustering)

A C++ implementation of the Silhouette method for interpreting and validating consistency within acca clusters of data.

Usage

sil_acca(acca, m, ...)

# S3 method for class 'acca_list'
sil_acca(acca, m, ...)

# S3 method for class 'list'
sil_acca(acca, m, ...)

Arguments

acca

[acca_list(1)]
Acca clustering results from acca.

m

[cmatrix(1)|matrix(1)]
A correlation matrix from corr_matrix. By default, the distance matrix (dist) used in this method is given by dist = 1 - m.

...

Additional arguments.

Value

[numeric(1)]
The average value of the silhouette width across all data in the entire dataset. Observations with a large average silhouette width (close to 1) are very well clustered.

References

Leonard Kaufman; Peter J. Rousseeuw (1990). Finding Groups in Data: An Introduction to Cluster Analysis. Hoboken, NJ: Wiley-Interscience. p. 87. doi:10.1002/9780470316801. ISBN 9780471878766.

Starczewski, Artur, and Adam Krzyżak. "Performance Evaluation of the Silhouette Index." International Conference on Artificial Intelligence and Soft Computing. Springer, Cham, 2015.

Author

Igor D.S. Siciliani, Paulo H. dos Santos

Examples

x <- corrp::corrp(iris)
m <- corrp::corr_matrix(x)
acca <- corrp::acca(m, 2)
sil_acca(acca, m)
#> [1] -0.2934118
#> attr(,"class")
#> [1] "corrpstat"
#> attr(,"statistic")
#> [1] "Silhouette"